requirements.txt
ultralytics
numpy
pandas
gdown>=3.10.1
Pillow>=5.2.0
opencv-python>=3.4.4
tensorflow>=1.9.0
retina-face
Pillow
dlib
matplotlib
tf-keras
torch
lightning
scikit-learn
optuna
gradio
torch
torchvision
lightning